pyproject.toml
[project]
name = "video-to-notion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.25.1",
    "azure-storage-blob>=12.27.1",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "yt-dlp>=2025.11.12",
]
